requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.7
kombu==5.3.7
multidict==6.0.5
mysqlclient==2.2.4
oauthlib==3.2.2
prompt_toolkit==3.0.47
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
twilio==9.2.3
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
